% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgse.R
\name{pgse}
\alias{pgse}
\title{Run PGSE}
\usage{
pgse(
  x,
  labels,
  pre_kfold_info_file = NULL,
  k = 6,
  ext = 2,
  target = 70,
  features = 10000,
  folds = 0,
  ea_min = NULL,
  ea_max = NULL,
  num_rounds = 1500,
  lr = 0.03,
  dist = FALSE,
  nodes = 1,
  workers = 1
)
}
\arguments{
\item{x}{character vector of file paths}

\item{labels}{vector of labels}

\item{pre_kfold_info_file}{path to pre-kfold info file \link{not supported yet}}

\item{k}{initial $k$-mer length}

\item{ext}{$p$ extension parameter}

\item{target}{target number of features}

\item{features}{number of features to select}

\item{folds}{number of folds for cross-validation (0 for train test split)}

\item{ea_min}{minimum essential agreement}

\item{ea_max}{maximum essential agreement}

\item{num_rounds}{number of rounds for XGBoost training}

\item{lr}{learning rate for XGBoost}

\item{dist}{logical, whether to use distributed training}

\item{nodes}{number of nodes for distributed training}

\item{workers}{number of workers for distributed training (or XGBoost)}
}
\value{
pgse_output_simple or pgse_output_cv object
}
\description{
Run PGSE
}
