% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pgse_output_cv}
\alias{plot.pgse_output_cv}
\title{Plot a cross-validation PGSE output}
\usage{
\method{plot}{pgse_output_cv}(x, y, ...)
}
\arguments{
\item{x}{pgse_output_cv object}

\item{y}{not used}

\item{...}{additional arguments}
}
\description{
Plot a cross-validation PGSE output
}
