% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pgse_output_simple}
\alias{plot.pgse_output_simple}
\title{Plot a simple PGSE output}
\usage{
\method{plot}{pgse_output_simple}(x, y, ...)
}
\arguments{
\item{x}{pgse_output_simple object}

\item{y}{not used}

\item{...}{additional arguments}
}
\description{
Plot a simple PGSE output
}
