% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{pgse_api_inference}
\alias{pgse_api_inference}
\title{Low-level API wrapper for PGSE inference pipeline}
\usage{
pgse_api_inference(model_path, segment_path, files, workers = 1L, ...)
}
\arguments{
\item{model_path}{path to the model file (usually .json)}

\item{segment_path}{path to the segments file (usually .txt)}

\item{files}{character vector of file paths to sequences}

\item{workers}{number of workers to use for inference}

\item{...}{additional arguments to pass to the pipeline}
}
\value{
array of predictions
}
\description{
This is low-level wrapper around the Python PGSE inference pipeline.
In general, it is recommended to use the \code{predict} method on
\code{pgse()} instead.
}
