% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgse.R
\name{pgse_api_train}
\alias{pgse_api_train}
\title{Low-level API for PGSE training pipeline}
\usage{
pgse_api_train(
  x,
  labels,
  pre_kfold_info_file = NULL,
  save_file = "",
  export_file = "./default.export",
  k = 6L,
  ext = 2L,
  target = 70L,
  features = 10000L,
  folds = 0L,
  ea_min = NULL,
  ea_max = NULL,
  num_rounds = 1500L,
  lr = 0.03,
  dist = FALSE,
  nodes = 1L,
  workers = 1L,
  ...
)
}
\arguments{
\item{x}{character vector of file paths}

\item{labels}{vector of labels}

\item{pre_kfold_info_file}{path to pre-kfold info file}

\item{save_file}{path to save file}

\item{export_file}{path to export file}

\item{k}{initial $k$-mer length}

\item{ext}{$p$ extension parameter}

\item{target}{target number of features}

\item{features}{number of features to select}

\item{folds}{number of folds for cross-validation (0 for train test split)}

\item{ea_min}{minimum essential agreement}

\item{ea_max}{maximum essential agreement}

\item{num_rounds}{number of rounds for XGBoost training}

\item{lr}{learning rate for XGBoost}

\item{dist}{logical, whether to use distributed training}

\item{nodes}{number of nodes for distributed training}

\item{workers}{number of workers for distributed training (or XGBoost)}

\item{...}{additional arguments to pass to the pipeline}
}
\value{
NULL (results are saved to disk)
}
\description{
Low-level API for PGSE training pipeline
}
