% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{predict.pgse_model}
\alias{predict.pgse_model}
\title{Predict method for PGSE model}
\usage{
\method{predict}{pgse_model}(object, newdata)
}
\arguments{
\item{object}{pgse_model object}

\item{newdata}{character list of file paths to sequences}
}
\value{
array of predictions
}
\description{
Predict method for PGSE model
}
