% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{predict.pgse_output_cv}
\alias{predict.pgse_output_cv}
\title{Predict method for cross-validation PGSE output}
\usage{
\method{predict}{pgse_output_cv}(object, newdata, workers = 1L, ...)
}
\arguments{
\item{object}{pgse_output_cv object}

\item{newdata}{character list of file paths to sequences}

\item{workers}{number of workers to use for inference}

\item{...}{additional arguments to pass to the pipeline}
}
\description{
Predict method for cross-validation PGSE output
}
