% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{predict.pgse_output_simple}
\alias{predict.pgse_output_simple}
\title{Predict method for PGSE output}
\usage{
\method{predict}{pgse_output_simple}(object, newdata, workers = 1L)
}
\arguments{
\item{object}{pgse_output object}

\item{newdata}{character list of file paths to sequences}

\item{workers}{number of workers to use for inference}
}
\value{
list of predictions (arrays)
}
\description{
Predict method for PGSE output
}
